# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with email and password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Exchange a valid refresh token for a new token pair",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalid or expired"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Revoke a refresh token / end a session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "summary": "(Bonus) Verify a user email using the token sent at registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified"
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "get": {
        "operationId": "AuthController_verifyEmailGet",
        "summary": "(Bonus) Verify a user email via query param token",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/chat/messages": {
      "post": {
        "operationId": "ChatController_sendMessage",
        "summary": "Send a prompt and receive an AI response",
        "description": "Sends a prompt to the selected (or default) AI provider and returns the full response in one shot. For an incrementally-streamed response, use POST /chat/messages/stream instead.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persisted user + assistant messages"
          },
          "400": {
            "description": "Validation failed on the request body"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Daily usage limit reached, or provider disabled"
          },
          "404": {
            "description": "No matching provider configured for this user"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/send": {
      "post": {
        "operationId": "ChatController_sendMessageAlias",
        "summary": "Send a prompt and receive an AI response (alias)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/messages/stream": {
      "post": {
        "operationId": "ChatController_sendMessageStream",
        "summary": "(Bonus) Send a prompt and stream the AI response as Server-Sent Events",
        "description": "Same behavior as POST /chat/messages, but the response is `text/event-stream`: a series of `event: chunk` frames as tokens arrive, followed by one `event: done` frame with the persisted message, or `event: error` if something went wrong.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "text/event-stream of chunk/done/error frames"
          },
          "400": {
            "description": "Validation failed on the request body"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Daily usage limit reached, or provider disabled"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "operationId": "ChatController_sendMessageStreamAlias",
        "summary": "(Bonus) Send prompt and stream AI response (alias)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/conversations": {
      "get": {
        "operationId": "ChatController_listConversations",
        "summary": "List the current user's conversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of conversations, most recent first"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/history": {
      "get": {
        "operationId": "ChatController_listHistoryAlias",
        "summary": "List the current user's conversations (alias)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/chat/conversations/{id}": {
      "get": {
        "operationId": "ChatController_getHistory",
        "summary": "Get full message history for a conversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation with its ordered messages"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Conversation not found or not owned by this user"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ChatController_deleteConversation",
        "summary": "Delete a conversation and its messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Conversation not found or not owned by this user"
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT Backend API",
    "description": "Production-ready backend for the EchoGPT Chrome Extension: auth, subscriptions, AI provider management, chat, web search, and admin analytics.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "example": "Summarize this page for me"
          },
          "message": {
            "type": "string",
            "example": "Summarize this page for me"
          },
          "content": {
            "type": "string",
            "example": "Summarize this page for me"
          },
          "conversationId": {
            "type": "string",
            "description": "Existing conversation to append to; omit to start a new one"
          },
          "providerId": {
            "type": "string",
            "description": "Specific provider id to use; omits to use the default provider"
          }
        }
      },
...
    }
  }
}